+++ src/test/java/study/cglib/HelloTarget.java
@@ -1,18 +1,15 @@
-package study.example;
+package study.cglib;
 
-public class HelloTarget implements Hello {
-	@Override
+public class HelloTarget {
 	public String sayHello(String name) {
 		return "Hello " + name;
 	}
 
-	@Override
 	public String sayHi(String name) {
 		return "Hi " + name;
 	}
 
-	@Override
 	public String sayThankYou(String name) {
-		return "Thank you " + name;
+		return "Thank You " + name;
 	}
 }
