+++ src/test/java/study/dynamicProxy/HelloTarget.java
@@ -1,15 +1,18 @@
-package study.example;
+package study.dynamicProxy;
 
-public class HelloTarget2 {
+public class HelloTarget implements Hello {
+	@Override
 	public String sayHello(String name) {
 		return "Hello " + name;
 	}
 
+	@Override
 	public String sayHi(String name) {
 		return "Hi " + name;
 	}
 
+	@Override
 	public String sayThankYou(String name) {
-		return "Thank You " + name;
+		return "Thank you " + name;
 	}
 }
